+++ server/app.js
@@ -22,7 +22,6 @@ const host = "easypeasy"
 const port = process.env.PORT || 8080
 const env = process.env.NODE_ENV || "development"
 const base = path.join(__dirname, "..")
-const views = path.join(base, "client/dist")
 
 /**
  * Locals
@@ -33,7 +32,6 @@ app.locals.host = host
 app.locals.port = port
 app.locals.env = env
 app.locals.base = base
-app.locals.views = views
 
 /**
  * Settings
@@ -44,18 +42,10 @@ app.disable("x-powered-by")
 app.set("json spaces", 2)
 
 /**
- * View engine
+ * Mount static assets
  */
 
-app.engine('jsx', express_react_views.createEngine())
-app.set('view engine', 'jsx')
-app.set("views", views)
-
-/**
- * Routes
- */
-
-app.use("/", require("./routes/root_router")(app))
+app.use("/", express.static(base + "/client/dist/assets"))
 
 /**
  * Start server